MyCrossword/server/pyproject.toml
[tool.poetry]
name = "mycrossword"
version = "0.1.0"
description = ""
authors = ["Seungwoo (Simon) Kim <seungwoo.simon.kim@gmail.com>", ]

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.0.5"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.13.0"
cryptography = "^37.0.2"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
